Moving tests to appropriate spec files and adding tests for new defined types
